+++ tests/testthat/test.stat_dist_slabinterval.R
@@ -232,3 +232,19 @@ test_that("distributional objects work", {
   )
 
 })
+
+test_that("stat_dist_ works on factor dist names", {
+  skip_if_not_installed("vdiffr")
+  skip_if_not_installed("svglite")
+
+  p = data.frame(
+    x = factor(c("norm", "norm")),
+    y = factor(c("a", "b"))
+  ) %>%
+    ggplot(aes(dist = x, y = y))
+
+  vdiffr::expect_doppelganger("stat_dist_ with factor dist name",
+    p + stat_dist_slabinterval()
+  )
+
+})